feat(ft_memset): implement

+++ ft_strlen.c
@@ -6,9 +6,11 @@
 /*   By: elagouch <elagouch@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 15:40:27 by elagouch          #+#    #+#             */
-/*   Updated: 2024/11/11 12:58:12 by elagouch         ###   ########.fr       */
+/*   Updated: 2024/11/11 13:25:49 by elagouch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "libft.h"
 
 /*
  * Gets the length of a string.
@@ -17,9 +19,9 @@
  *
  * @retuns	length of str
  */
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
